requirements.txt
Pillow==10.2.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117